+++ templates/admin-panel/template/client/__tests__/unit/components/App.js
@@ -1,4 +1,4 @@
-import App from '../../../src/components/App.sht';
+import App from '../../../src/components/App.sfc';
 import renderer from 'sham-ui-test-helpers';
 
 it( 'renders correctly', () => {