+++ templates/admin-panel/template/client/src/initializers/main.js
@@ -3,28 +3,20 @@ import hrefto from 'sham-ui-router/href-to';
 import Session from '../services/session';
 import Store from '../services/store';
 import startRouter from './routes';
-import App from '../components/App.sht';
+import App from '../components/App.sfc';
 
 export default function() {
     new Session();
-    const store = new Store();
+    new Store();
 
     startRouter();
 
-    const app = new App( {
+    new App( {
         ID: 'app',
         container: document.querySelector( 'body' ),
         directives: {
             ...directives,
             hrefto
-        },
-        tokenLoaded: false,
-        routerResolved: false
+        }
     } );
-
-    store.tokenPromise.then(
-        () => app.update( {
-            tokenLoaded: true
-        } )
-    );
 }