+++ templates/admin-panel/template/client/src/storages/app.js
@@ -1,0 +1,9 @@
+import createStorage from 'sham-ui-data-storage';
+
+export const { useStorage } = createStorage( {
+    routerResolved: false,
+    tokenLoaded: false
+}, {
+    DI: 'app:storage',
+    sync: true
+} );